refactor: add not-authorized check
