+++ pages/home/index.tsx
@@ -1,9 +1,24 @@
 import { getLayout } from '@/components/Layout/Layout'
 import { PageWrapper } from '@/components/PageWrapper/PageWrapper'
+import { authRouts } from '@/components/common/Auth/authRoutes'
 import { userRouts } from '@/components/common/User/userRouts'
+import { checkAuth } from '@/utils/checkAuth'
 import { useRouter } from 'next/router'
 
-//add me-Query in getServerSideProps
+export const getServerSideProps = async () => {
+  if (!checkAuth()) {
+    return {
+      redirect: {
+        destination: authRouts.notAuthorized,
+        permanent: false,
+      },
+    }
+  }
+
+  return {
+    props: {},
+  }
+}
 
 const Home = () => {
   const router = useRouter()
